refacto: use an autoroute
