--- packages/trpc/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@betagouv/hyyypertool...trpc",
-  "version": "0.0.0-alpha.1",
-  "module": "index.ts",
-  "type": "module",
-  "dependencies": {
-    "@trpc/server": "^10.44.1",
-    "zod": "^3.22.4"
-  }
-}
