+++ src/assets/config.ts
@@ -1,0 +1,7 @@
+//
+
+import env from ":common/env";
+
+//
+
+export const ASSETS_PATH = `/assets/${env.VERSION}` as const;
