+++ src/assets/rewrite.ts
@@ -1,0 +1,9 @@
+//
+
+import { ASSETS_PATH } from "./config";
+
+//
+
+export function rewriteAssetRequestPath(path: string) {
+  return path.replace(ASSETS_PATH, "");
+}
