+++ src/assets/tailwind.css
