+++ src/legacy/routes/moderations/:id/index.tsx
@@ -1,0 +1,19 @@
+//
+
+import { Id_Schema } from ":common/schema";
+import { ModerationPage } from ":legacy/moderations/page";
+import { zValidator } from "@hono/zod-validator";
+import { Hono } from "hono";
+import { renderToReadableStream } from "hono/jsx/streaming";
+
+//
+
+export default new Hono().get(
+  "/",
+  zValidator("param", Id_Schema),
+  async ({ body, req }) => {
+    const { id } = req.valid("param");
+
+    return body(renderToReadableStream(<ModerationPage active_id={id} />));
+  },
+);
